split up everything and removed duplicated functionality
